added main menu, credits, game over scenes
game over is not finished being programmed, since I don't have universal player names. If you'd like to scrap the scene because that won't be easy to do, I can create a popup window when the game ends to do that instead. I just found myself without a lot of the necessary information to code certain things:/
How to connect main menu:
the scripts should already be set up, but you've got to adjust how it builds
open main menu scene
go to file > build settings
click to add current scene
do this for all the scenes
Reorder so that Main Menu: 0, Game: 1, Game Over: 2, Credits: 3
Or I'll do it in a little while.